Updating tests -mocked logger -renamed some Tests to ITs
