+++ src/test/java/com/safetynet/alerts/presentation/controller/GetMappingControllerIT.java
@@ -1,11 +1,14 @@
 package com.safetynet.alerts.presentation.controller;
 
 import com.safetynet.alerts.data.io.JsonDAO;
+import com.safetynet.alerts.logging.LogHandlerTiny;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.web.WebAppConfiguration;
@@ -16,10 +19,10 @@ import org.springframework.web.context.WebApplicationContext;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@WebMvcTest(MedicalRecordController.class)
+@WebMvcTest(GetMappingController.class)
 @WebAppConfiguration
-@Import(ObjectControllerBeanSetup.class)
-public class MedicalRecordControllerTest {
+@Import(GetControllerBeanSetup.class)
+public class GetMappingControllerIT {
 
     @Autowired
     private MockMvc mvc;
@@ -106,12 +109,12 @@ public class MedicalRecordControllerTest {
 
     //Testing successful interactions
     @Test
-    public void medicalRecordEndpointCanAddNewMedicalRecordViaPost() throws Exception {
+    public void firestationReturnsPeopleServicedByStationNumber() throws Exception {
         //Preparation
-        String uri = "/medicalRecord?firstName=Testame&lastName=Name&birthDate=01/01/2000&medications=something:1g,another:2g&allergies=nothing,really";
+        String uri = "/firestation?stationNumber=1";
 
         //Method
-        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.post(uri)
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                 .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
 
         //Verification
@@ -119,80 +122,260 @@ public class MedicalRecordControllerTest {
         String receivedResponse = mvcResult.getResponse().getContentAsString()
                 .replace("\n", "").replace("  ", "");
 
-        assertEquals(201, status);
-        assertEquals(TestConstants.medicalRecordControllerPostExpectedResponse,receivedResponse);
+        assertEquals(200, status);
+        assertEquals(TestConstants.firestationReturnsPeopleServicedByStationNumber,receivedResponse);
     }
 
     @Test
-    public void medicalRecordEndpointCanUpdateMedicalRecordViaPut() throws Exception {
+    public void childAlertReturnsListOfChildrenAtAddress() throws Exception {
         //Preparation
-        String uri = "/medicalRecord?firstName=John&lastName=Boyd&birthDate=01/01/2000&medications=&allergies=";
+        String uri = "/childAlert?address=1509 Culver St";
 
         //Method
-        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.put(uri)
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                 .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
-        int status = mvcResult.getResponse().getStatus();
 
         //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
         assertEquals(200, status);
+        assertEquals(TestConstants.childAlertReturnsListOfChildrenAtAddress,receivedResponse);
     }
 
     @Test
-    public void medicalRecordEndpointCanRemoveMedicalRecordViaDelete() throws Exception {
+    public void phoneAlertReturnsListOfPhoneNumbersForFirestationNumber() throws Exception {
         //Preparation
-        String uri = "/medicalRecord?firstName=John&lastName=Boyd";
+        String uri = "/phoneAlert?firestation=1";
 
         //Method
-        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.delete(uri)
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                 .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
-        int status = mvcResult.getResponse().getStatus();
 
         //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
         assertEquals(200, status);
+        assertEquals(TestConstants.phoneAlertReturnsListOfPhoneNumbersForFirestationNumber,receivedResponse);
+    }
+
+    @Test
+    public void fireReturnsFirestationNumberAndResidentsForAddress() throws Exception {
+        //Preparation
+        String uri = "/fire?address=1509 Culver St";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.fireReturnsFirestationNumberAndResidentsForAddress,receivedResponse);
+    }
+
+    @Test
+    public void floodStationsReturnsListOfHouseholdsForEachProvidedStationNumber() throws Exception {
+        //Preparation
+        String uri = "/flood/stations?stations=2,4";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.floodStationsReturnsListOfHouseholdsForEachProvidedStationNumber,receivedResponse);
+    }
+
+    @Test
+    public void floodStationsReturnsListOfHouseholdsForEachProvidedStationNumberEvenIfSomeHaveNoResidents() throws Exception {
+        //Preparation
+        String uri = "/flood/stations?stations=1,10";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.floodStationsReturnsListOfHouseholdsForEachProvidedStationNumberEvenIfSomeHaveNoResidents,receivedResponse);
+    }
+
+    @Test
+    public void personInfoReturnsDetailsForAllPeopleWithProvidedName() throws Exception {
+        //Preparation
+        String uri = "/personInfo?firstName=John&lastName=Boyd";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.personInfoReturnsDetailsForAllPeopleWithProvidedName,receivedResponse);
+    }
+
+    @Test
+    public void communityEmailReturnsEmailAddressesForAllResidentsOfProvidedCity() throws Exception {
+        //Preparation
+        String uri = "/communityEmail?city=Culver";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.communityEmailReturnsEmailAddressesForAllResidentsOfProvidedCity,receivedResponse);
     }
 
     //Testing unsuccessful interactions
     @Test
-    public void medicalRecordEndpointReturns409WhenPostingDuplicateMedicalRecord() throws Exception {
+    public void firestationReturnsEmptyJsonIfStationHasNoResidentsAssociated() throws Exception {
         //Preparation
-        String uri = "/medicalRecord?firstName=John&lastName=Boyd&birthDate=01/01/2000&medications=something:1g,another:2g&allergies=nothing,really";
+        String uri = "/firestation?stationNumber=10";
 
         //Method
-        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.post(uri)
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                 .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
 
         //Verification
         int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
 
-        assertEquals(409, status);
+        assertEquals(200, status);
+        assertEquals(TestConstants.firestationEmptyJson,receivedResponse);
     }
 
     @Test
-    public void medicalRecordEndpointReturnsError404WhenUpdatingNonExistingMedicalRecord() throws Exception {
+    public void childAlertReturnsEmptyStringIfNoChildrenAtAddress() throws Exception {
         //Preparation
-        String uri = "/medicalRecord?firstName=Doesnt&lastName=Exist&birthDate=01/01/2000&medications=&allergies=";
+        String uri = "/childAlert?address=908 73rd St";
 
         //Method
-        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.put(uri)
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                 .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
-        int status = mvcResult.getResponse().getStatus();
 
         //Verification
-        assertEquals(404, status);
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals("",receivedResponse);
     }
 
     @Test
-    public void medicalRecordEndpointReturnsError404WhenRemovingNonExistingMedicalRecord() throws Exception {
+    public void phoneAlertReturnsEmptyJsonIfFirestationNumberHasNoResidentPhoneNumbersAssociated() throws Exception {
         //Preparation
-        String uri = "/medicalRecord?firstName=Doesnt&lastName=Exist&birthDate=01/01/2000&medications=&allergies=";
+        String uri = "/phoneAlert?firestation=10";
 
         //Method
-        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.delete(uri)
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                 .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
-        int status = mvcResult.getResponse().getStatus();
 
         //Verification
-        assertEquals(404, status);
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.phoneAlertEmptyJson,receivedResponse);
     }
 
+    @Test
+    public void fireReturnsEmptyJsonIfAddressIsNotFound() throws Exception {
+        //Preparation
+        String uri = "/fire?address=1 Fake St";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.fireEmptyJson,receivedResponse);
+    }
+
+    @Test
+    public void floodStationsReturnsEmptyJsonIfNoProvidedStationNumberHasResidents() throws Exception {
+        //Preparation
+        String uri = "/flood/stations?stations=10,11";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.floodStationsEmptyJson,receivedResponse);
+    }
+
+    @Test
+    public void personInfoReturnsEmptyJsonIfNoPeopleWithProvidedName() throws Exception {
+        //Preparation
+        String uri = "/personInfo?firstName=Doesnt&lastName=Exist";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.personInfoEmptyJson,receivedResponse);
+    }
+
+    @Test
+    public void communityEmailReturnsEmptyJsonIfNoResidentsRegisteredForProvidedCity() throws Exception {
+        //Preparation
+        String uri = "/communityEmail?city=Denver";
+
+        //Method
+        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
+                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();
+
+        //Verification
+        int status = mvcResult.getResponse().getStatus();
+        String receivedResponse = mvcResult.getResponse().getContentAsString()
+                .replace("\n", "").replace("  ", "");
+
+        assertEquals(200, status);
+        assertEquals(TestConstants.communityEmailEmptyJson,receivedResponse);
+    }
 }
